Add files via upload
The features used for analysis were extrapolated from the main features to make informed decisions. The target variable was the level of claims of a car (low, medium, high). Six classification algorithms were used to make predictions. The results were not the best due to a high imbalance in the target variable.